Simple-IOS-Profile-Downloader.txt
This payload goes out to a URL (Hopefully where there is a IOS profile prompt) and downloads the profile. Then navigates to VPN & Device Management where the profile can be fully installed by the user

Some timing on this payload may have to be changed depending on the device and its version. This payload was tested on iOS 16.1 but should work most newer versions of iOS

In order to use a O.MG device on a iPhone you must have a Lighting to USB-A camera adapter

Make sure to specify the URL within payload before use