Create payload.txt
This payload adds a ActiveSync Microsoft exchange server to the iPhone giving the exchange server access to information such as Mail, Contacts, Calendars, Reminders, and Notes. The exchange server can also remotely wipe the device. This payloads timing will have to be changed in order to run against a device successfully to prevent wrong use, the payload has to be finished by the user accepting the exchange server connection. This is the manual config option of the payload.